mow-1365: Avoid duplicate undo point, store only selection, double vs int issues

+++ common/src/main/java/org/geogebra/common/euclidian/DrawableList.java
@@ -35,7 +35,7 @@ public class DrawableList extends ArrayList<Drawable> {
 		this.comparator = new Comparator<Drawable>() {
 			@Override
 			public int compare(Drawable a, Drawable b) {
-				return (int) comparator.compare(a.geo, b.geo, false);
+				return comparator.compare(a.geo, b.geo, false);
 			}
 		};
 	}
